Updating the handling of arrays to handle multiple puts into an inner array before it is added to the outer array for the first time.
Adding the ability to construct traces that go via put operations into
outer arrays.

These required or triggered the following updates:

* Tidying traces implementations and sorting an issue with only looking
at the scope of the sample, and not the enclosing scopes when working
out where to split the trace that writes back.

* Tightening a guard of traces so it catches implicit puts that appear
with other implicit puts in between.

* Updating the detection of IRTrees that write the a value back on
itself. i.e. a = a;

* And doing the same addition for the ability to track sets of possible
puts to an array.

* Correcting a guard on the trace comparison that was not correct as it
assumed all traces end in a sample task. This is no longer true.