Changing behaviour when variables are fixed so that the retention policy is not over written.
Adding in the behaviour where both the samples and the map value can be
saved from a models execution. This could be made more efficient by just
saving the index in the samples that the map occurred at, but this is
simpler and will do until the custom behaviour is required for space
reasons.