A number of corrections including: * Fixing a bug with the get length code that meant that earlier element lengths were being dropped when new arrays were added to an outer array.
* Fixed a bug where traces were grouped incorrectly if there was a set
which had no traces in common with the trace being added a new set
would be created even if this trace could and therefore should just be
added to a different set. A flag to track this case resolves this
problem.

* Moving scopes and a wrapper class to the correct packages.

* Combining the flags to disable parallel compilation of sandwood models
so that they only appear in a single location. This flag is only used
for debugging.

* Modified inference generator base to ensure that the code to update
sample values cannot span other generated code.