Several improvements to inference code that marginalises or experiments with proposed variables such as Metropolis Hastings.
* Modifications to probability inference used for marginalisation and MH
code so that the number of states is declared as a variable and then
this number is used both to iterate through the states to test, and for
normalising the stored results in the case of marginalisation.

* Adding code to explore all the possible number of states if an input
to a random variable is a distribution.

* Changing the behaviour of scope constructor so that values are passed
for samples that have not changed too as it could still be that the
trace reaches variables that were not reached by the earlier traces. The
optimiser will have to remove any unused temporary variables.

This code still has errors around the determination of the length of
arrays, and a rewrite will be required once the construction of
conditionals is complete so that array.length is only directly queried
for inputs, in all other cases a trace restriction is used to determine
the value of the original creation of the array.