Modifications so that lengths are now calculated by back tracking through the program rather than by just keeping a set of possible lengths. A set of possible lengths is still calculated as this is required to prevent infinite loops with distributed sampling. These values are now calculated on demand rather than on the fly, but are cached.
This approach is far more robust than the previous approach and is made
possible by the back tracking work done for conditionals. It also means
that arrays no longer have to be or are guaranteed to be the same length
as the underlying Java array. In future this should allow better memory
reuse via shared scratch space. It is also an important step towards
running on GPUs.