Modifying the variables so that the variables that are fixed by observations are tracked. This allows errors to be detected where contradictions could be introduced, and for working out which intermediate variables should not be updated by methods such as set intermediates.
Other fixes in this commit include:
Making sure errors generated by super classes are collected.

Fixing a bug in set alias if puts occur before it is called. This
happens in the construction of multi-dimensional arrays.

Modification so that observed data is not updated by calls to update
intermediates.

Removing models that are now rejected by the compiler as they could
introduce contradictions

Making observations visible from all array instances and preventing
direct access to the constraint parameter.

Adding a check to ensure that observed arrays are not modified.