A set of changes effecting the multinomial distribution, including:
Adding a check to confirm that the arguments to a multinomial random
variable are suitably constrained.

Propagating observed information through the multinomial, specifically
how many samples are taken.

Adding a guard so that inference can be skipped if the arguments to the
random variable mean it is not relevant.

Removing tests that were added to exercise the code generation, but are
now barred by the new guards.

Updating the basic test to the latest model.

Adding a new test where the count from an observed multinomial is
generated from a random variable.