* Adding a guard to prevent variables being declared twice. There is an issue here if the value is used in several conditions in a trace restriction in different iterations of a loop. Some thought will be required into how to handle this.
* Adding block scopes to ensure that there are not collisions on
declared variables.

* Renaming methods to make the code clearer.

* Changing the visibility of a constructor as it should never be used
outside of the class.

* Modifications to prevent reordering in the hash maps/sets. Updating
hash and equals methods.

* Modifications to the tracking of scopes, so that instead of having a
map of scopes to representative data structures where every scope maps
to a single point data structure representing the node the map is
updated as scopes are revisited on different iterations of for loops.
This modification is tied in with the storage of which variables are
initialised, so this can also change with loop iterations, while not
forgetting the variable names are are currently in scope. In the process
of doing this some other changes were made including:

* Preventing if else scopes being created if one of these is substituted
as this can only happen if this point was reached when constructing a
trace restriction.

* Removing calls to push scope state and substitutions except when
moving to generate code for a different function, for example when
producing a snippet of code that will work as an allocator.

* Fixing a bug where comments were added repeatedly.

* Modifying the generation of errors when trying to reduce
distributions.

* Removing code that was creating anonymous variables for generated
arguments as this generated code is no longer required due to the better
handling of initialisation recording.